requirements.txt
python-dotenv==1.0.1
gradio
google-generativeai
pylint
pytest==8.3.2
fire
black
pytest-cov